Add temporary change of splash screen
Expo SDK 52 introduced a weird thing with expo-splash-screen. Now, you can't use a full image as the splash screen, only an icon. This is problematic because my splash screen had my icon, app title, and phrase. In a few updates, this must be checked to see if they've better dealt with all the problems.